feat(stories): split SC & CC
