+++ components/stories/sp.ts
@@ -1,0 +1,5 @@
+export type SearchParams = {
+    series: string
+    s: string
+    c: string
+}
